Migrate docs for terra-alert and terra-arrange into terra-core-docs
