+++ packages/terra-core-docs/src/terra-dev-site/doc/alert/Alert.1.doc.mdx
@@ -1,18 +1,18 @@
 import { Badge } from 'terra-alert/package.json?dev-site-package';
 
-import AlertExample from '../example/AlertExample?dev-site-example';
-import ErrorExample from '../example/ErrorExample?dev-site-example';
-import WarningExample from '../example/WarningExample?dev-site-example';
-import AdvisoryExample from '../example/AdvisoryExample?dev-site-example';
-import UnsatisfiedExample from '../example/UnsatisfiedExample?dev-site-example';
-import UnverifiedExample from '../example/UnverifiedExample?dev-site-example';
-import InfoExample from '../example/InfoExample?dev-site-example';
-import SuccessExample from '../example/SuccessExample?dev-site-example';
-import CustomExample from '../example/CustomExample?dev-site-example';
-import LongTextExample from '../example/LongTextExample?dev-site-example';
-import ActionExample from '../example/ActionExample?dev-site-example';
-import DismissibleExample from '../example/DismissibleExample?dev-site-example';
-import ActionAndDismissibleExample from '../example/ActionAndDismissibleExample?dev-site-example';
+import AlertExample from './example/AlertExample?dev-site-example';
+import ErrorExample from './example/ErrorExample?dev-site-example';
+import WarningExample from './example/WarningExample?dev-site-example';
+import AdvisoryExample from './example/AdvisoryExample?dev-site-example';
+import UnsatisfiedExample from './example/UnsatisfiedExample?dev-site-example';
+import UnverifiedExample from './example/UnverifiedExample?dev-site-example';
+import InfoExample from './example/InfoExample?dev-site-example';
+import SuccessExample from './example/SuccessExample?dev-site-example';
+import CustomExample from './example/CustomExample?dev-site-example';
+import LongTextExample from './example/LongTextExample?dev-site-example';
+import ActionExample from './example/ActionExample?dev-site-example';
+import DismissibleExample from './example/DismissibleExample?dev-site-example';
+import ActionAndDismissibleExample from './example/ActionAndDismissibleExample?dev-site-example';
 
 import AlertPropsTable from 'terra-alert/src/Alert?dev-site-props-table';
