+++ packages/terra-core-docs/src/terra-dev-site/doc/alert/example/DismissibleExample.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import Button from 'terra-button';
-import Alert from '../../../Alert';
+import Alert from 'terra-alert';
 
 const AlertDismissible = () => {
   const [isOpen, setIsOpen] = useState(true);
