+++ packages/terra-core-docs/src/terra-dev-site/doc/alert/example/InfoExample.jsx
