+++ packages/terra-core-docs/src/terra-dev-site/doc/alert/example/LongTextExample.module.scss
