+++ packages/terra-core-docs/src/terra-dev-site/doc/arrange/Arrange.1.doc.mdx
@@ -1,9 +1,9 @@
 import { Badge } from 'terra-arrange/package.json?dev-site-package';
 
-import ArrangeAlignAllContainers from '../example/ArrangeAlignAllContainers?dev-site-example';
-import ArrangeAlignFill from '../example/ArrangeAlignFill?dev-site-example';
-import ArrangeAlignFitEnd from '../example/ArrangeAlignFitEnd?dev-site-example';
-import ArrangeAlignFitStart from '../example/ArrangeAlignFitStart?dev-site-example';
+import ArrangeAlignAllContainers from './example/ArrangeAlignAllContainers?dev-site-example';
+import ArrangeAlignFill from './example/ArrangeAlignFill?dev-site-example';
+import ArrangeAlignFitEnd from './example/ArrangeAlignFitEnd?dev-site-example';
+import ArrangeAlignFitStart from './example/ArrangeAlignFitStart?dev-site-example';
 
 import ArrangePropsTable from 'terra-arrange/src/Arrange?dev-site-props-table';
