+++ packages/terra-core-docs/src/terra-dev-site/doc/arrange/common/examplesetup.jsx
