+++ packages/terra-core-docs/src/terra-dev-site/doc/arrange/example/ArrangeAlignAllContainers.jsx
