+++ packages/terra-core-docs/src/terra-dev-site/test/alert/AlertTitle.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Alert from '../../../Alert';
+import Alert from 'terra-alert';
 
 export default () => (
   <div>
